Feature #2767 changed a load more floats to double

+++ Code/Gui/MITK/HelperApps/niftkProjectTrackedPointsOnStereoVideo.cxx
@@ -70,20 +70,20 @@ int main(int argc, char** argv)
     projector->SetTrackerIndex(trackerIndex);
     projector->SetDrawAxes(DrawAxes);
     
-    std::vector < std::pair < cv::Point2f, cv::Point2f > > screenPoints;
+    std::vector < std::pair < cv::Point2d, cv::Point2d > > screenPoints;
     unsigned int setPointsFrameNumber;
-    std::vector < cv::Point3f > worldPoints;
+    std::vector < cv::Point3d > worldPoints;
     if ( input2D.length() != 0 ) 
     {
       std::ifstream fin(input2D.c_str());
       fin >> setPointsFrameNumber;
-      float x1;
-      float y1;
-      float x2;
-      float y2;
+      double x1;
+      double y1;
+      double x2;
+      double y2;
       while ( fin >> x1 >> y1 >> x2 >> y2 )
       {
-        screenPoints.push_back(std::pair<cv::Point2f,cv::Point2f> (cv::Point2f(x1,y1), cv::Point2f(x2,y2)));
+        screenPoints.push_back(std::pair<cv::Point2d,cv::Point2d> (cv::Point2d(x1,y1), cv::Point2d(x2,y2)));
       }
       fin.close();
       projector->SetWorldPointsByTriangulation(screenPoints,setPointsFrameNumber);
@@ -91,12 +91,12 @@ int main(int argc, char** argv)
   if ( input3D.length() != 0 ) 
     {
       std::ifstream fin(input3D.c_str());
-      float x;
-      float y;
-      float z;
+      double x;
+      double y;
+      double z;
       while ( fin >> x >> y >> z  )
       {
-        worldPoints.push_back(cv::Point3f(x,y,z));
+        worldPoints.push_back(cv::Point3d(x,y,z));
       }
       projector->SetWorldPoints(worldPoints);
       fin.close();
@@ -107,7 +107,7 @@ int main(int argc, char** argv)
     if ( output2D.length() != 0 ) 
     {
       std::ofstream fout (output2D.c_str());
-      std::vector < std::vector < std::pair < cv::Point2f , cv::Point2f > > > projectedPoints = 
+      std::vector < std::vector < std::pair < cv::Point2d , cv::Point2d > > > projectedPoints = 
         projector->GetProjectedPoints();
       fout << "#Frame Number " ;
       for ( unsigned int i = 0 ; i < projectedPoints[0].size() ; i ++ ) 
@@ -130,7 +130,7 @@ int main(int argc, char** argv)
     if ( output3D.length() !=0 )
     {
       std::ofstream fout (output3D.c_str());
-      std::vector < std::vector < cv::Point3f > > leftLensPoints = 
+      std::vector < std::vector < cv::Point3d > > leftLensPoints = 
         projector->GetPointsInLeftLensCS();
       fout << "#Frame Number " ;
       for ( unsigned int i = 0 ; i < leftLensPoints[0].size() ; i ++ )
