+++ Code/Gui/MITK/Modules/niftkOpenCV/VideoTrackerMatching/mitkProjectPointsOnStereoVideo.h
@@ -71,14 +71,14 @@ public:
   /**
    * \brief Set the world points directly
    */
- // void SetWorldPoints (std::vector<cv::Point3f> worldPoints);
+ // void SetWorldPoints (std::vector<cv::Point3d> worldPoints);
 
   /**
    * \brief Set the world points by triangulating their position from the
    * on screen coordinates for the specified frame
    */
   void SetWorldPointsByTriangulation 
-    (std::vector< std::pair<cv::Point2f,cv::Point2f> > onScreenPointPairs, 
+    (std::vector< std::pair<cv::Point2d,cv::Point2d> > onScreenPointPairs, 
      unsigned int FrameNumber);
 
   void SetVisualise( bool) ;
@@ -86,9 +86,9 @@ public:
   itkSetMacro ( TrackerIndex, int);
   itkSetMacro ( DrawLines, bool);
   itkSetMacro ( DrawAxes, bool);
-  itkSetMacro ( WorldPoints, std::vector<cv::Point3f> );
-  std::vector < std::vector <cv::Point3f> > GetPointsInLeftLensCS();
-  std::vector < std::vector < std::pair<cv::Point2f, cv::Point2f> > >  GetProjectedPoints();
+  itkSetMacro ( WorldPoints, std::vector<cv::Point3d> );
+  std::vector < std::vector <cv::Point3d> > GetPointsInLeftLensCS();
+  std::vector < std::vector < std::pair<cv::Point2d, cv::Point2d> > >  GetProjectedPoints();
   itkGetMacro ( InitOK, bool);
   itkGetMacro ( ProjectOK, bool);
 
@@ -118,7 +118,7 @@ private:
   std::string                   m_VideoIn; //the video in file
   std::string                   m_VideoOut; //video needs to be saved on the fly
   std::string                   m_Directory; //the directory containing the data
-  std::vector<cv::Point3f>      m_WorldPoints;  //the world points to project
+  std::vector<cv::Point3d>      m_WorldPoints;  //the world points to project
 
   int                           m_TrackerIndex; //the tracker index to use for frame matching
   mitk::VideoTrackerMatching::Pointer
@@ -138,11 +138,11 @@ private:
   cv::Mat* m_RightToLeftTranslationVector;
   cv::Mat* m_LeftCameraToTracker;
 
-  std::vector < std::vector < std::pair<cv::Point2f, cv::Point2f> > >
+  std::vector < std::vector < std::pair<cv::Point2d, cv::Point2d> > >
                                 m_ProjectedPoints; // the projected points
-  std::vector < std::vector <cv::Point3f> >    
+  std::vector < std::vector <cv::Point3d> >    
                                 m_PointsInLeftLensCS; // the points in left lens coordinates.
-  std::vector < std::pair<cv::Point2f, cv::Point2f> > 
+  std::vector < std::pair<cv::Point2d, cv::Point2d> > 
                                 m_ScreenAxesPoints; // the projected axes points
 
   CvCapture*                    m_Capture;
